backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
dj-database-url==2.1.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
Pillow==10.0.1
drf-spectacular==0.26.5
python-decouple==3.8
django-taggit==4.0.0
django-extensions==3.2.3
gunicorn==21.2.0 
